Bring sidebar into own component and bring filter settings into general object
